# docs/openapi/gateway.openapi.json
{
  "components": {
    "schemas": {
      "ErrorResponse": {
        "example": {
          "detail": "Authentication required. Provide Authorization: Bearer <token>."
        },
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowRun": {
        "description": "Full workflow run record. Must be persisted and resumable from checkpoint.",
        "example": {
          "checkpoint": {
            "uploaded_files": 3
          },
          "created_at": "2026-04-05T12:35:00Z",
          "current_step": "extract_rfis",
          "project_id": "proj_01hv6n8h0q9x2r4s6t8u1v3w5y",
          "status": "running",
          "steps": [
            {
              "checkpoint": {
                "uploaded_files": 3
              },
              "completed_at": "2026-04-05T12:38:00Z",
              "started_at": "2026-04-05T12:35:00Z",
              "status": "complete",
              "step_name": "ingest_documents"
            }
          ],
          "tenant_id": "tenant_01hv6n6j6h8x7m4r9k2p1q3s4t",
          "type": "project_intake",
          "updated_at": "2026-04-05T12:40:00Z",
          "workflow_id": "wf_01hv6r7v6j8x7m4r9k2p1q3s4t"
        },
        "properties": {
          "checkpoint": {
            "additionalProperties": true,
            "title": "Checkpoint",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_step": {
            "title": "Current Step",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "type",
          "tenant_id",
          "project_id",
          "status",
          "current_step",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowRun",
        "type": "object"
      },
      "WorkflowStartRequest": {
        "example": {
          "inputs": {
            "source_files": [
              "gs://robco-staging-bucket/intake/project-brief.pdf",
              "gs://robco-staging-bucket/intake/rfis.xlsx"
            ]
          },
          "project_id": "proj_01hv6n8h0q9x2r4s6t8u1v3w5y",
          "tenant_id": "tenant_01hv6n6j6h8x7m4r9k2p1q3s4t",
          "workflow_type": "project_intake"
        },
        "properties": {
          "inputs": {
            "additionalProperties": true,
            "title": "Inputs",
            "type": "object"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "workflow_type": {
            "title": "Workflow Type",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "project_id",
          "workflow_type"
        ],
        "title": "WorkflowStartRequest",
        "type": "object"
      },
      "WorkflowStatus": {
        "enum": [
          "running",
          "complete",
          "failed",
          "paused"
        ],
        "title": "WorkflowStatus",
        "type": "string"
      },
      "WorkflowStep": {
        "description": "Individual step record within a workflow run.",
        "example": {
          "checkpoint": {
            "processed_pages": 12
          },
          "started_at": "2026-04-05T12:40:00Z",
          "status": "running",
          "step_name": "extract_rfis"
        },
        "properties": {
          "checkpoint": {
            "additionalProperties": true,
            "title": "Checkpoint",
            "type": "object"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "step_name": {
            "title": "Step Name",
            "type": "string"
          }
        },
        "required": [
          "step_name",
          "status"
        ],
        "title": "WorkflowStep",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "description": "Bearer token returned by POST /v1/auth/token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "robcofamily@gmail.com",
      "name": "Robco Platform"
    },
    "description": "Centralized inference control plane and policy engine for tenant-scoped authentication, normalized inference, workflow entry, and control-plane tenant administration.",
    "title": "Robco Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workflows/start": {
      "post": {
        "description": "Creates an initial workflow run and hands off execution to orchestration.",
        "operationId": "start_workflow_v1_workflows_start_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Project-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/{workflow_id}": {
      "get": {
        "description": "Looks up workflow status by workflow ID.",
        "operationId": "get_workflow_v1_workflows__workflow_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Project-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Workflow ID format is invalid."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Workflow was not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get workflow status",
        "tags": [
          "workflows"
        ]
      }
    }
  }
}